fix(lint): resolve all eslint errors introduced by new features
- tokenizer: remove unnecessary \! escapes in sheet-prefix regexes
- cross-sheet.test: merge duplicate imports from types and cross-sheet-types
- workflow-form: extract condition-field builders to workflow-form-condition-fields.ts (max-lines)
- trigger-condition-evaluator.test: move nested/empty compound tests to companion file (max-lines)
